+++ packages/components/src/components/bal-date/bal-date-calendar/bal-date-calendar.tsx
@@ -40,6 +40,8 @@ import { CalendarNav } from './components/bal-date-calendar__nav'
 export class DateCalendar implements ComponentInterface, Loggable, BalConfigObserver, BalSwipeObserver {
   private yearListEl: HTMLUListElement | undefined
   private gridEl: HTMLDivElement | undefined
+  private oldMin? = ''
+  private oldMax? = ''
 
   @Element() el!: HTMLElement
 
@@ -122,6 +124,7 @@ export class DateCalendar implements ComponentInterface, Loggable, BalConfigObse
   rangePropChanged() {
     this.months = generateMonths(this.language, this.year, this.selectedDate, this.min, this.max)
     this.years = generateYears(this.year, this.minYear, this.maxYear)
+    this.generateGrid()
   }
 
   /**
@@ -223,6 +226,15 @@ export class DateCalendar implements ComponentInterface, Loggable, BalConfigObse
    * ------------------------------------------------------
    */
 
+  private generateGrid() {
+    const date = BalDate.fromISO(this.value)
+    if (date.isValid) {
+      this.generateGridByDate(date)
+    } else {
+      this.generateGridByDefaultDate()
+    }
+  }
+
   private generateGridByDate(date: BalDate) {
     if (date.year !== undefined && date.month !== undefined) {
       this.generateGridByMonthAndYear(date.month, date.year)
@@ -245,12 +257,14 @@ export class DateCalendar implements ComponentInterface, Loggable, BalConfigObse
 
   private generateGridByMonthAndYear(month: number, year: number) {
     // Only generate when mont and year has changed
-    if (this.month !== month || this.year !== year) {
+    if (this.month !== month || this.year !== year || this.oldMin !== this.min || this.oldMax !== this.max) {
       this.calendarGrid = generateCalendarGrid(year, month, this.min, this.max, this.allowedDates)
       this.firstDayOfWeek = getFirstWeekdayOfMonth(year, month)
     }
     this.month = month
     this.year = year
+    this.oldMin = this.min
+    this.oldMax = this.max
   }
 
   /**